Resolve the control-flow tutorial conflict with #3271 by keeping its
general untraced-field paragraph and pointing the enum section at it.

Agent-Harness: Claude Code
Agent-Model: claude-opus-5-5[1m]
Agent-Session: https://claude.ai/code/session_01592Hc7Y1ZjeqdXirDGww44

+++ src/accelerators/Registration.jl
@@ -23,7 +23,7 @@ end
 
 function register_backend(
     platform_name::String;
-    priority::Int64,
+    priority::Int,
     pjrt_initialize_function=nothing,
     ifrt_initialize_function=nothing,
     preinitialize_setup_function=Returns(nothing),
@@ -43,8 +43,8 @@ function register_backend(
             priority,
             pjrt_initialize_function,
             ifrt_initialize_function,
-            Ref(0),
-            Ref(0),
+            Ref(Int64(0)),
+            Ref(Int64(0)),
             preinitialize_setup_function,
         )
         push!(RegisteredBackends, backend)
